fix(imu): address CodeRabbit review findings (round 12)
- Use smallest typed enum underlying types (uint8_t for timing, uint16_t for stack size)
- Add @test tags to rx_bmp280_init and rx_bmp280_read Doxygen
- Add second @pre/@post to all test functions in test_rx_bno055.c
- Replace magic numbers with named constants in invariant comments
- Remove redundant (press_comp_constants_t) cast in compensate_pressure
- Add rx_log_error call when BMP280 status register read fails
- Add clarifying comment to post-condition RX_ASSERTs in rx_bno055_test_reset_state

+++ star-rx72n-firmware/src/tasks/inc/imu_task.h
@@ -93,7 +93,7 @@ typedef enum : uint8_t {
  *
  * @since Version 1.0.0
  */
-typedef enum : uint16_t {
+typedef enum : uint8_t {
   k_imu_task_period_ms = 50U, /**< IMU sampling period in milliseconds (20 Hz) */
   k_imu_task_period_ticks =
     5U, /**< IMU sampling period in RTOS ticks (5 ticks x 10 ms/tick = 50 ms at 100 Hz) */
@@ -112,7 +112,7 @@ typedef enum : uint16_t {
  *
  * @since Version 1.0.0
  */
-typedef enum : uint32_t {
+typedef enum : uint16_t {
   k_imu_task_stack_size_bytes = 2048U, /**< Stack size in bytes for IMU task thread */
 } imu_task_stack_t;
